Added spinner on item click listener
