+++ core/src/main/java/in/hangang/core/view/appbar/BaseAppBarClickListener.kt
@@ -2,7 +2,7 @@ package `in`.hangang.core.view.appbar
 
 import android.view.View
 
-interface HangangAppBarClickListener {
+interface BaseAppBarClickListener {
     fun onBackButtonClick(v: View)
     fun onRightButtonClick(v: View)
 }
