+++ core/src/main/java/in/hangang/core/view/button/checkbox/FilledCheckBox.kt
@@ -1,16 +1,17 @@
-package `in`.hangang.core.view.button.radiobutton
+package `in`.hangang.core.view.button.checkbox
 
 import `in`.hangang.core.R
 import android.content.Context
 import android.util.AttributeSet
 import android.view.Gravity
+import android.widget.CheckBox
 import androidx.core.content.ContextCompat
 
-class HangangFilledRadioButton @JvmOverloads constructor(
+class FilledCheckBox @JvmOverloads constructor(
     context: Context,
     attributeSet: AttributeSet? = null,
     defStyleAttr: Int = 0
-) : androidx.appcompat.widget.AppCompatRadioButton(context, attributeSet, defStyleAttr) {
+) : androidx.appcompat.widget.AppCompatCheckBox(context, attributeSet, defStyleAttr) {
 
     init {
         minWidth = resources.getDimensionPixelSize(R.dimen.radio_min_width)
