+++ core/src/main/java/in/hangang/core/view/spinner/BottomSheetSpinner.kt
@@ -5,19 +5,18 @@ import android.content.Context
 import android.util.AttributeSet
 import android.view.LayoutInflater
 import android.view.View
-import android.widget.ArrayAdapter
-import android.widget.LinearLayout
-import android.widget.ListView
-import android.widget.TextView
+import android.widget.*
 import com.google.android.material.bottomsheet.BottomSheetDialog
 
-class HangangBottomSheetSpinner @JvmOverloads constructor(
-    context: Context,
-    attributeSet: AttributeSet? = null,
-    defStyleAttr: Int = 0,
-    defStyleRes: Int = 0
+class BottomSheetSpinner @JvmOverloads constructor(
+        context: Context,
+        attributeSet: AttributeSet? = null,
+        defStyleAttr: Int = 0,
+        defStyleRes: Int = 0
 ) : LinearLayout(context, attributeSet, defStyleAttr, defStyleRes) {
     val view: View = LayoutInflater.from(context).inflate(R.layout.layout_spinner, this, true)
+
+    var onItemClickListener: AdapterView.OnItemClickListener? = null
 
     var position = -1
         set(value) {
@@ -35,7 +34,6 @@ class HangangBottomSheetSpinner @JvmOverloads constructor(
     private var _textViewSelectedItem: TextView
 
     val textViewSelectedItem: TextView get() = _textViewSelectedItem
-    val selection get() = _textViewSelectedItem.text
 
     init {
         _textViewSelectedItem = view.findViewById(R.id.text_view_spinner_selected_item)
@@ -43,18 +41,32 @@ class HangangBottomSheetSpinner @JvmOverloads constructor(
         this.setOnClickListener {
             showBottomSheet()
         }
+
+        context.theme.obtainStyledAttributes(
+                attributeSet,
+                R.styleable.HangangBottomSheetSpinner,
+                defStyleAttr,
+                defStyleRes
+        ).apply {
+            val entries = getTextArray(R.styleable.HangangBottomSheetSpinner_android_entries)
+            if(entries != null) {
+                items = entries.toList()
+            }
+        }
     }
 
     private fun showBottomSheet() {
         val bottomSheet = BottomSheetDialog(context, R.style.HangangBottomSheetDialogTheme)
         val bottomSheetView =
-            LayoutInflater.from(context).inflate(R.layout.layout_spinner_bottom_sheet, null)
+                LayoutInflater.from(context).inflate(R.layout.layout_spinner_bottom_sheet, null)
         bottomSheetView.findViewById<ListView>(R.id.spinner_bottom_sheet_list_view).apply {
             adapter =
-                ArrayAdapter(context, R.layout.simple_text_item_1, android.R.id.text1, items)
+                    ArrayAdapter(context, R.layout.simple_text_item_1, android.R.id.text1, items)
             divider = null
-            setOnItemClickListener { _, _, position, _ ->
-                this@HangangBottomSheetSpinner.position = position
+
+            setOnItemClickListener { parent, view, position, id ->
+                this@BottomSheetSpinner.position = position
+                this@BottomSheetSpinner.onItemClickListener?.onItemClick(parent, view, position, id)
                 bottomSheet.dismiss()
             }
         }
