+++ core/src/main/res/values/attrs.xml
@@ -18,4 +18,8 @@
         <attr name="max" format="integer|reference" />
         <attr name="progress" format="integer|reference" />
     </declare-styleable>
+
+    <declare-styleable name="HangangBottomSheetSpinner">
+        <attr name="android:entries" />
+    </declare-styleable>
 </resources>
